Rename interactive components to block elements
