+++ src/block-kit/index.ts
@@ -11,8 +11,8 @@ export { Image } from './Image'
 export { Input } from './Input'
 export { Section, Field } from './Section'
 
-// Interactive components
-export { Button } from './interactive/Button'
+// Block elements
+export { Button } from './elements/Button'
 export {
   Select,
   SelectFragment,
@@ -22,9 +22,9 @@ export {
   UsersSelect,
   ConversationsSelect,
   ChannelsSelect,
-} from './interactive/Select'
-export { Overflow, OverflowItem } from './interactive/Overflow'
-export { DatePicker } from './interactive/DatePicker'
+} from './elements/Select'
+export { Overflow, OverflowItem } from './elements/Overflow'
+export { DatePicker } from './elements/DatePicker'
 
 // Composition objects
 export { Confirm } from './composition/Confirm'
